+++ src/main/java/com/teamharmony/newscommunity/users/controller/UserController.java
@@ -5,9 +5,9 @@ import com.auth0.jwt.JWTVerifier;
 import com.auth0.jwt.algorithms.Algorithm;
 import com.auth0.jwt.interfaces.DecodedJWT;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.teamharmony.newscommunity.users.dto.response.UserResponseDto;
+import com.teamharmony.newscommunity.users.dto.UserResponseDto;
 import com.teamharmony.newscommunity.users.vo.ProfileVO;
-import com.teamharmony.newscommunity.users.dto.request.SignupRequestDto;
+import com.teamharmony.newscommunity.users.dto.SignupRequestDto;
 import com.teamharmony.newscommunity.users.entity.*;
 import com.teamharmony.newscommunity.users.service.UserService;
 import lombok.Data;
