+++ src/main/java/com/teamharmony/newscommunity/users/dto/ProfileResponseDto.java
@@ -1,4 +1,4 @@
-package com.teamharmony.newscommunity.users.dto.response;
+package com.teamharmony.newscommunity.users.dto;
 
 import com.teamharmony.newscommunity.users.entity.UserProfile;
 import lombok.AllArgsConstructor;
